Remove owncast
couldn't get it working. I think I was having a firewall issue but
couldn't figure it on the server or client side. OBS would crash and
owncast would never see the incoming rtmp stream. It seems like the
issue is a firewall rule on the server... but can't figure out what was
block owncast...

Signed-off-by: Taylor Silva <taylorsilva@outlook.com>